+++ corePlugins/mcFunction/stringReader.py
@@ -1,17 +1,14 @@
 import re
+from dataclasses import dataclass, field
 from typing import Optional, final
 
+from base.model.parsing.parser import _Base
 from cat.utils.collections_ import Stack
-from base.model.parsing.bytesUtils import DIGITS, ASCII_LETTERS, JAVA_WHITESPACES, JAVA_WHITESPACES_SINGLE_BYTE, JAVA_WHITESPACES_THREE_BYTES, DIGITS_RANGE
+from base.model.parsing.bytesUtils import DIGITS, ASCII_LETTERS, JAVA_WHITESPACES, JAVA_WHITESPACES_SINGLE_BYTE, JAVA_WHITESPACES_THREE_BYTES, DIGITS_RANGE, ORD_BACKSLASH
 from base.model.utils import Position, Span
 
 Char = bytes
 Byte = int
-
-
-SYNTAX_ESCAPE = '\\'
-SYNTAX_DOUBLE_QUOTE = '"'
-SYNTAX_SINGLE_QUOTE = '\''
 
 
 DOT_OR_MINUS = set(b'.-')
@@ -23,46 +20,34 @@ NOT_JAVA_WHITESPACES_REGEX: bytes = b"(?:" + b'|'.join(JAVA_WHITESPACES) + b")*"
 
 
 @final
-class StringReader:
-	def __init__(self, source: bytes, lineStart: int, lineNo: int, cursorOffset: int, fullSource: bytes):
-		self._lineStart: int = lineStart
-		self._lineNo: int = lineNo
-		self.source: bytes = source
-		self.totalLength = len(source)
-		self.cursor: int = 0
-		self.cursorOffset: int = cursorOffset
-		self.lastCursors: Stack[int] = Stack()
-		self.fullSource: bytes = fullSource
+@dataclass
+class StringReader(_Base):
+	fullSource: bytes
+	lastCursors: Stack[int] = field(default_factory=Stack, init=False)
 
 	@property
 	def hasReachedEnd(self) -> bool:
-		return self.cursor >= self.totalLength
+		return self.cursor >= self.length
 
 	def posFromColumn(self, cursor: int) -> Position:
-		actualCursor = (cursor + self._lineStart) + self.cursorOffset
-		return Position(self._lineNo, actualCursor - self._lineStart, actualCursor)
-		# return Position(self._lineNo, cursor, self._lineStart + cursor)
-
-	@property
-	def currentPos(self) -> Position:
-		return self.posFromColumn(self.cursor)
+		# ugh. Definitely not threadsafe, but it gets the job done.
+		currentCursor = self.cursor
+		self.cursor = cursor
+		pos = self.currentPos
+		self.cursor = currentCursor
+		return pos
 
 	@property
 	def currentSpan(self) -> Span:
 		start = self.lastCursors.peek()
 		begin = self.posFromColumn(start)
-		end = self.posFromColumn(self.cursor)
+		end = self.currentPos
 		return Span(begin, end)
 
 	def save(self) -> None:
 		self.lastCursors.push(self.cursor)
 
 	def mergeLastSave(self) -> None:
-		""" removes the last save, without resetting the cursor. """
-		assert self.lastCursors
-		self.lastCursors.pop()
-
-	def acceptLastSave(self) -> None:
 		""" removes the last save, without resetting the cursor. """
 		assert self.lastCursors
 		self.lastCursors.pop()
@@ -79,13 +64,13 @@ class StringReader:
 
 	def tryConsumeWhitespace(self) -> bool:
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 		# beware of unicode utf-8:
 		while cursor < length:
-			if source[cursor] in JAVA_WHITESPACES_SINGLE_BYTE:
+			if text[cursor] in JAVA_WHITESPACES_SINGLE_BYTE:
 				cursor += 1
-			elif cursor + 3 <= length and source[cursor:cursor + 3] in JAVA_WHITESPACES_THREE_BYTES:
+			elif cursor + 3 <= length and text[cursor:cursor + 3] in JAVA_WHITESPACES_THREE_BYTES:
 				cursor += 3
 			else:
 				break
@@ -96,22 +81,15 @@ class StringReader:
 		else:
 			return False
 
-	# @cy_final
-	# def tryConsumeChar(self, char: Char) -> bool:
-	# 	if self.cursor < self.totalLength and self.source[self.cursor] == char:
-	# 		self.cursor += 1
-	# 		return True
-	# 	return False
-
 	def tryConsumeByte(self, byte: Byte) -> bool:
-		if self.cursor < self.totalLength and self.source[self.cursor] == byte:
+		if self.cursor < self.length and self.text[self.cursor] == byte:
 			self.cursor += 1
 			return True
 		return False
 
 	def tryPeek(self) -> Optional[Byte]:
-		if self.cursor < self.totalLength:
-			return self.source[self.cursor]
+		if self.cursor < self.length:
+			return self.text[self.cursor]
 		else:
 			return None
 
@@ -120,31 +98,31 @@ class StringReader:
 			return None
 		start: int = self.cursor
 		self.save()
-		self.cursor = self.totalLength
-		return self.source[start:]
+		self.cursor = self.length
+		return self.text[start:]
 
 	def readUntilEndOr(self, sub: bytes, *, includeTerminator: bool = False) -> bytes:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 
-		cursor = source.find(sub, cursor)
+		cursor = text.find(sub, cursor)
 		if cursor == -1:
 			cursor = length
 		elif includeTerminator:
 			cursor += len(sub)
 		self.save()
 		self.cursor = cursor
-		return source[start:cursor]
+		return text[start:cursor]
 
 	def readUntilEndOrRegex(self, sub: re.Pattern[bytes], *, includeTerminator: bool = False) -> bytes:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 
-		findIter = sub.finditer(source, cursor)
+		findIter = sub.finditer(text, cursor)
 		match = next(findIter, None)
 		if match is None:
 			cursor = length
@@ -155,7 +133,7 @@ class StringReader:
 
 		self.save()
 		self.cursor = cursor
-		return source[start:cursor]
+		return text[start:cursor]
 
 	def readUntilEndOrWhitespace(self) -> bytes:
 		result = self.tryReadRegex(re.compile(NOT_JAVA_WHITESPACES_REGEX))
@@ -163,7 +141,7 @@ class StringReader:
 		return result
 
 	def tryReadRegex(self, pattern: re.Pattern[bytes]) -> Optional[bytes]:  # throws CommandSyntaxException
-		match = pattern.match(self.source, self.cursor)
+		match = pattern.match(self.text, self.cursor)
 		if match is None:
 			return None
 		text = match.group(0)
@@ -174,85 +152,85 @@ class StringReader:
 	def tryReadInt(self) -> Optional[bytes]:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 
-		if cursor < length and source[cursor] == ord('-'):
+		if cursor < length and text[cursor] == ord('-'):
 			cursor += 1
-		if cursor < length and source[cursor] in DIGITS_RANGE:
+		if cursor < length and text[cursor] in DIGITS_RANGE:
 			cursor += 1
 		else:
 			return None
-		while cursor < length and source[cursor] in DIGITS_RANGE:
+		while cursor < length and text[cursor] in DIGITS_RANGE:
 			cursor += 1
-		if cursor < length and source[cursor] in DOT_OR_MINUS:
+		if cursor < length and text[cursor] in DOT_OR_MINUS:
 			return None  # TODO: Lexer Errors
 		self.save()
 		self.cursor = cursor
-		return source[start:cursor]
+		return text[start:cursor]
 
 	def tryReadFloat(self) -> Optional[bytes]:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 		hasHadDot: bool = False
 
-		if cursor < length and source[cursor] == ord('-'):
+		if cursor < length and text[cursor] == ord('-'):
 			cursor += 1
 
-		if cursor < length and source[cursor] == ord('.'):
+		if cursor < length and text[cursor] == ord('.'):
 			hasHadDot = True
 			cursor += 1
 
-		if cursor < length and source[cursor] in DIGITS_RANGE:
+		if cursor < length and text[cursor] in DIGITS_RANGE:
 			cursor += 1
 		else:
 			return None
-		while cursor < length and source[cursor] in DIGITS_RANGE:
+		while cursor < length and text[cursor] in DIGITS_RANGE:
 			cursor += 1
-		if cursor < length and source[cursor] == ord('.'):
+		if cursor < length and text[cursor] == ord('.'):
 			cursor += 1
 			if hasHadDot:
 				return None  # TODO: Lexer Errors
-		while cursor < length and source[cursor] in DIGITS_RANGE:
+		while cursor < length and text[cursor] in DIGITS_RANGE:
 			cursor += 1
-		if cursor < length and source[cursor] in DOT_OR_MINUS:
+		if cursor < length and text[cursor] in DOT_OR_MINUS:
 			return None  # TODO: Lexer Errors
 		self.save()
 		self.cursor = cursor
-		return source[start:cursor]
+		return text[start:cursor]
 
 	def tryReadString(self) -> Optional[bytes]:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
+		text: bytes = self.text
+		length: int = self.length
 		if cursor >= length:
 			return None
-		terminator: Byte = source[cursor]
-		TERMINATOR_OR_ESCAPE: set[Byte] = {terminator, ord(SYNTAX_ESCAPE)}
+		terminator: Byte = text[cursor]
+		TERMINATOR_OR_ESCAPE: set[Byte] = {terminator, ORD_BACKSLASH}
 
 		if terminator in QUOTES:  # quoted string
 			cursor += 1
 			strStreakStart = cursor
 			result: bytes = b''
 			while cursor < length:
-				c: Byte = source[cursor]
+				c: Byte = text[cursor]
 
-				if c == ord(SYNTAX_ESCAPE):
-					result += source[strStreakStart:cursor]
+				if c == ORD_BACKSLASH:
+					result += text[strStreakStart:cursor]
 					cursor += 1
 					if not cursor < length:
 						return None
-					c2 = source[cursor]
+					c2 = text[cursor]
 					if c2 in TERMINATOR_OR_ESCAPE:
 						strStreakStart = cursor
 						cursor += 1
 					else:
 						return None  # TODO: Lexer Errors
 				elif c == terminator:
-					result += source[strStreakStart:cursor]
+					result += text[strStreakStart:cursor]
 					cursor += 1
 					self.save()
 					self.cursor = cursor
@@ -262,54 +240,13 @@ class StringReader:
 
 			return None  # TODO: Lexer Errors
 		else:  # unquoted string
-			while cursor < length and source[cursor] in UNQUOTED_STRING_CHARS:
+			while cursor < length and text[cursor] in UNQUOTED_STRING_CHARS:
 				cursor += 1
 			if start == cursor:
 				return None
 			self.save()
 			self.cursor = cursor
-			return source[start:cursor]
-
-	# @cy_final
-	# def tryReadString(self) -> Optional[bytes]:
-	# 	start: int = self.cursor
-	# 	cursor: int = self.cursor
-	# 	source: bytes = self.source
-	# 	length: int = self.totalLength
-	# 	if cursor >= length:
-	# 		return None
-	# 	terminator: Byte = source[cursor]
-	# 	TERMINATOR_OR_ESCAPE: set[Byte] = {terminator, ord(SYNTAX_ESCAPE)}
-	# 	if terminator in QUOTES:  # quoted string
-	# 		cursor += 1
-	# 		result: bytes = b''
-	# 		escaped: bool = False
-	# 		while cursor < length:
-	# 			c: Byte = source[cursor]
-	# 			cursor += 1
-	# 			if escaped:
-	# 				if c in TERMINATOR_OR_ESCAPE:
-	# 					result += c
-	# 					escaped = False
-	# 				else:
-	# 					return None  # TODO: Lexer Errors
-	# 			elif c == ord(SYNTAX_ESCAPE):
-	# 				escaped = True
-	# 			elif c == terminator:
-	# 				self.save()
-	# 				self.cursor = cursor
-	# 				return result
-	# 			else:
-	# 				result += c
-	# 		return None  # TODO: Lexer Errors
-	# 	else: # unquoted string
-	# 		while cursor < length and source[cursor] in UNQUOTED_STRING_CHARS:
-	# 			cursor += 1
-	# 		if start == cursor:
-	# 			return None
-	# 		self.save()
-	# 		self.cursor = cursor
-	# 		return source[start:cursor]
+			return text[start:cursor]
 
 	def tryReadBoolean(self) -> Optional[bytes]:  # throws CommandSyntaxException
 		value: Optional[bytes] = self.tryReadString()
@@ -324,75 +261,36 @@ class StringReader:
 	def tryReadLiteral(self) -> Optional[bytes]:  # throws CommandSyntaxException
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
-		if cursor >= length or source[cursor] == ord(' '):
+		text: bytes = self.text
+		length: int = self.length
+		if cursor >= length or text[cursor] == ord(' '):
 			return None
-		while cursor < length and source[cursor] != ord(' '):
+		while cursor < length and text[cursor] != ord(' '):
 			cursor += 1
 		self.save()
 		self.cursor = cursor
-		return source[start:cursor]
-
-	# @cy_final
-	# def tryReadEntityLocator(self) -> Optional[str]:
-	# 	# Must be a player name, a target selector or a UUID.
-	# 	start: int = self.cursor
-	# 	cursor: int = self.cursor
-	# 	source: str = self.source
-	# 	length: int = self.totalLength
-	#
-	# 	string = self.tryReadString()
-	# 	if string is None:
-	# 		# tryParse Target selector:
-	# 		if cursor + 1 < length and source[cursor:cursor + 2] in {'@p', '@r', '@a', '@e', '@s'}:
-	# 			cursor += 2
-	# 			if cursor < length and source[cursor] == '[':  # No spaces before '['!!
-	# 				cursor2: int = source.find(']', cursor + 1)
-	# 				if cursor2 != -1:
-	# 					cursor = cursor2 + 1
-	# 			self.save()
-	# 			self.cursor = cursor
-	# 			return source[start:cursor]
-	# 		else:
-	# 			return None
-	# 	else:
-	# 		return string
+		return text[start:cursor]
 
 	def tryReadTildeNotation(self) -> Optional[bytes]:
-		start: int = self.cursor
-		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
-		if cursor < length and source[cursor] == ord('~'):
-			cursor += 1
-			self.save()
-			self.cursor = cursor
-			if self.tryReadFloat() is not None:
-				self.mergeLastSave()
-			return source[start:self.cursor]
-			# if cursor == length or source[cursor] == ' ':
-			# 	return source[start:self.cursor]
-			# else:
-			# 	self.rollback()
-			# 	return None
-		else:
-			return None
+		return self._tryReadNotation(ord('~'))
 
 	def tryReadCaretNotation(self) -> Optional[bytes]:
+		return self._tryReadNotation(ord('^'))
+
+	def _tryReadNotation(self, notation: int) -> Optional[bytes]:
 		start: int = self.cursor
 		cursor: int = self.cursor
-		source: bytes = self.source
-		length: int = self.totalLength
-		if cursor < length and source[cursor] == ord('^'):
+		text: bytes = self.text
+		length: int = self.length
+		if cursor < length and text[cursor] == notation:
 			cursor += 1
 			self.save()
 			self.cursor = cursor
 			if self.tryReadFloat() is not None:
 				self.mergeLastSave()
-			return source[start:self.cursor]
-			# if cursor == length or source[cursor] == ' ':
-			# 	return source[start:self.cursor]
+			return text[start:self.cursor]
+			# if cursor == length or text[cursor] == ' ':
+			# 	return text[start:self.cursor]
 			# else:
 			# 	self.rollback()
 			# 	return None