+++ corePlugins/mcFunction/utils.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from base.model.utils import Message, GeneralError
 
-EXPECTED_ARGUMENT_SEPARATOR_MSG: Message = Message("Expected whitespace to end one argument, but found trailing data: `{}`", 1)
-
 
 class CommandSyntaxError(GeneralError):
 	pass
@@ -14,7 +12,6 @@ class CommandSemanticsError(GeneralError):
 
 
 __all__ = [
-	'EXPECTED_ARGUMENT_SEPARATOR_MSG',
 	'CommandSyntaxError',
 	'CommandSemanticsError',
 ]